+++ packages/xrpl/src/client/index.ts
@@ -48,7 +48,7 @@ import type {
 } from '../models/methods/subscribe'
 import type { SubmittableTransaction } from '../models/transactions'
 import { areAmountsEqual } from '../models/transactions/common'
-import { setTransactionFlagsToNumber } from '../models/utils/flags'
+import { convertTxFlagsToNumber } from '../models/utils/flags'
 import {
   ensureClassicAddress,
   submitRequest,
@@ -666,7 +666,7 @@ class Client extends EventEmitter<EventTypes> {
     const tx = { ...transaction }
 
     setValidAddresses(tx)
-    setTransactionFlagsToNumber(tx)
+    tx.Flags = convertTxFlagsToNumber(tx)
 
     const promises: Array<Promise<void>> = []
     if (tx.NetworkID == null) {