+++ packages/xrpl/HISTORY.md
@@ -2,12 +2,14 @@
 
 Subscribe to [the **xrpl-announce** mailing list](https://groups.google.com/g/xrpl-announce) for release announcements. We recommend that xrpl.js (ripple-lib) users stay up-to-date with the latest stable release.
 
-## Unreleased Changes
+## Unreleased
 
 ### Added
+* Adds utility function `convertTxFlagsToNumber`
 * New util `areAmountsEqual` to check if 2 amounts are strictly equal
 
-### Fixed
+### Changed
+* Deprecated `setTransactionFlagsToNumber`. Start using convertTxFlagsToNumber instead
 * `autofill` function in client not validating amounts correctly
 
 ## 4.1.0 (2024-12-23)